#7 - Bug fix for recursiveLoan application creation & Package restructering.
Fixed the bug for recursive loan application creation in LoanApplicationCreationFlow.java when the CA responds with score check processing.
Restructered the packages based on the node parties.
