+++ workflows/src/main/java/com/xyz/flows/bank/BankLoanDisbursementFlow.java
@@ -1,4 +1,4 @@
-package com.xyz.flows;
+package com.xyz.flows.bank;
 
 public class BankLoanDisbursementFlow {
 }
