+++ workflows/src/main/java/com/xyz/observer/ca/CACreditScoreCheckStateObserver.java
@@ -1,7 +1,7 @@
-package com.xyz.observer;
+package com.xyz.observer.ca;
 
 import com.xyz.constants.CreditScoreDesc;
-import com.xyz.flows.CreditCheckProcessingFlow;
+import com.xyz.flows.ca.CreditCheckProcessingFlow;
 import com.xyz.states.CreditRatingState;
 import net.corda.core.contracts.UniqueIdentifier;
 import net.corda.core.identity.CordaX500Name;
@@ -16,13 +16,13 @@ import rx.Observable;
 
 import java.util.concurrent.ExecutionException;
 
-public class CreditCheckInitiationObserver {
-    private static final Logger logger = LoggerFactory.getLogger(CreditCheckInitiationObserver.class);
+public class CACreditScoreCheckStateObserver {
+    private static final Logger logger = LoggerFactory.getLogger(CACreditScoreCheckStateObserver.class);
 
     private final CordaRPCOps proxy;
     private final CordaX500Name me;
 
-    public CreditCheckInitiationObserver(CordaRPCOps proxy, CordaX500Name name) {
+    public CACreditScoreCheckStateObserver(CordaRPCOps proxy, CordaX500Name name) {
         this.proxy = proxy;
         this.me = name;
     }
@@ -41,8 +41,7 @@ public class CreditCheckInitiationObserver {
                     final UniqueIdentifier creditCheckApplicationId = creditCheckState.getLoanVerificationId();
 
                     logger.info("Initiating Credit Check flow from observer for CreditCheck Application ID  : " + creditCheckApplicationId);
-                    Thread newThreadCreditCheckFlow = new Thread(new InitiateCreditCheckProcess(creditCheckApplicationId, proxy));
-                    newThreadCreditCheckFlow.start();
+                    new CACreditScoreCheckStateTrigger(creditCheckApplicationId, proxy).trigger();
                 }
             }));
         } catch (Exception e) {
@@ -52,29 +51,31 @@ public class CreditCheckInitiationObserver {
     }
 }
 
-class InitiateCreditCheckProcess implements Runnable {
-    private static final Logger logger = LoggerFactory.getLogger(InitiateCreditCheckProcess.class);
+class CACreditScoreCheckStateTrigger {
+    private static final Logger logger = LoggerFactory.getLogger(CACreditScoreCheckStateTrigger.class);
 
     private final UniqueIdentifier creditCheckApplicationId;
     private final CordaRPCOps proxy;
 
-    public InitiateCreditCheckProcess(UniqueIdentifier creditCheckApplicationId, CordaRPCOps proxy) {
+    public CACreditScoreCheckStateTrigger(UniqueIdentifier creditCheckApplicationId, CordaRPCOps proxy) {
         this.creditCheckApplicationId = creditCheckApplicationId;
         this.proxy = proxy;
     }
 
-    @Override
-    public void run() {
+    public void trigger() {
         try {
-            logger.info("INitiating the credit check flow but sleeping for 2 minutes from now");
+            logger.info("========###########################======");
+            logger.info("INTENTIONAL SLEEP OF 2 MINUTES, TO VERIFY STATES INVOKING APIs");
             Thread.sleep(120000);
-            logger.info("Starting CreditScoreCheckFlow for Loan ApplicationID : " + this.creditCheckApplicationId.getId().toString());
-
+            logger.info("Starting processing for CreditScore evaluation for CreditCheck ApplicationID : " + this.creditCheckApplicationId.getId().toString());
             Party financeAgencyNode = proxy.wellKnownPartyFromX500Name(CordaX500Name.parse("O=XYZLoaning,L=London,C=GB"));
-            SignedTransaction tx = proxy.startTrackedFlowDynamic(CreditCheckProcessingFlow.class, this.creditCheckApplicationId, financeAgencyNode).getReturnValue().get();
-            logger.info("Credit Check flow updated with CreditCheck Application Id : " +
-                    ((CreditRatingState) tx.getTx().getOutputs().get(0).getData()).getLoanVerificationId().toString());
 
+            SignedTransaction tx = proxy.startTrackedFlowDynamic(CreditCheckProcessingFlow.class, this.creditCheckApplicationId, financeAgencyNode).getReturnValue().get();
+            CreditRatingState caState = ((CreditRatingState) tx.getTx().getOutputs().get(0).getData());
+
+            logger.info("Credit Check flow updated with CreditCheck Application Id : " +
+                    caState.getLoanVerificationId().toString() + " With Status : " + caState.getCreditScoreDesc().toString() + " and Score : " + caState.getCreditScoreCheckRating());
+            logger.info("=================##########===============");
         } catch (InterruptedException | ExecutionException e) {
             logger.error("Error while initiating CreditScoreCheckFlow for loanApplicationId : " + creditCheckApplicationId.getId().toString());
             e.printStackTrace();
